requirements.txt
outline-vpn-api
python-decouple
sqlalchemy
aiohttp
aiogram
